Texture mapping added
Created a new vukan image and a view that goes into that image as well as device memory that is bound to the image.
Added UV coordinates to the vertex input and altered  code in the shaders to take advantage of these coordinates.
Added a sampler and a descriptor that links the sample to the image view.