Depth Buffer Added
Created a depth buffer "image" along with the memory and a view into the image.
The image was added as an attachment to the only rendering pass, and was added as an attachment during framebuffer creation.
Finally, the depth buffer was enabled and configured using a state structure and passed to the pipeline create info.